refactor: make separate repo for versions #37
